mailbox: qcom-apcs-ipc: add IPQ8074 APSS clock controller support
IPQ8074 has the APSS clock controller utilizing the same register space as
the APCS, so provide access to the APSS utilizing a child device like
IPQ6018 does as well, but just by utilizing the IPQ8074 specific APSS
clock driver.

Also, APCS register space in IPQ8074 is 0x6000 so max_register needs to be
updated to 0x5ffc.

Signed-off-by: Robert Marko <robimarko@gmail.com>
---
Changes in v5:
* Use lower case hex for max_register
* Update the APSS clock name to match the new one without commas