arm64: dts: ipq8074: add APCS support
APCS now has support for providing the APSS clocks as the child device
for IPQ8074.

So, add the required DT node for APCS and A53PLL which feeds it the PLL
clock for APSS, since APSS will be used for CPU frequency scaling.

Signed-off-by: Robert Marko <robimarko@gmail.com>
---
Changes in v7:
* Drop review tag due to significant changes
* Add A53PLL node
* Add clocks to APCS node

Changes in v3:
* Node does not currently exist in the upstream kernel, so add it instead
of modifying.