arm64: dts: ipq8074: drop standalone SMEM node
Now that the SMEM binding and driver allows the SMEM node to be
described in the reserved-memory region directly, move the compatible
and hwlock properties to the reserved-memory node and drop the
standadlone node.

While at it update the TCSR as well to match the updated binding and
driver, this allows getting rid of syscon.

Signed-off-by: Robert Marko <robimarko@gmail.com>