arm64: dts: qcom: cp01-c1: add jedec,spi-nor compatible
Add the generic "jedec,spi-nor" compatible for SPI-NOR as compatible.
"n25q128a11" is a leftover from the vendor SDK, and boards ship with
various manufacturers and models of SPI-NOR (Mine uses mx25u6435f).

The "n25q128a11" compatible is left as the bootloader uses it for finding
the SPI-NOR to patch in the fixed-partitions before booting.

This gets rid of the following:
[    0.698067] spi-nor spi0.0: found mx25u6435f, expected n25q128a11

Signed-off-by: Robert Marko <robimarko@gmail.com>