clk: qcom: ipq6018: add QUP6 I2C clock
QUP6 I2C clock despite having an index in dt-bindings is not actually
implemented in the clock driver, looks like an oversight as remaining
QUP I2C clocks are implemented.

So, implement the clock in order to enable QUP6 I2C controller in DTS.

Note that it looks that it should not be disabled as for some reason
with DVFS it will cause the board to hang, this is what QCA does in their
downstream 5.4 kernel as well.

Signed-off-by: Robert Marko <robimarko@gmail.com>