PCI: Drop controller CONFIG_OF dependencies
Many drivers depend on OF interfaces, so they won't be functional if
CONFIG_OF is not set.  But OF provides stub functions in that case, so drop
the OF dependencies so we can at least compile-test the drivers.

This means we can compile-test the following drivers even without CONFIG_OF
enabled (many still require either CONFIG_COMPILE_TEST or the relevant
arch):

  aardvark
  al
  apple
  brcmstb
  cadence-host
  cadence-platform-host
  cadence-ep
  dra7xx-host
  dra7xx-ep
  dw-rockchip
  ftpci100
  hisi
  intel-gw
  ixp4xx
  j721e-host
  j721e-ep
  kirin
  layerscape-host
  layerscape-ep
  mediatek
  microchip-host
  mobiveil
  mvebu
  qcom-host
  qcom-ep
  rockchip-host
  rockchip-ep
  rockchip-dwc
  pci-host-generic
  uniphier-host
  uniphier-ep
  v3-semi
  xilinx

Link: https://lore.kernel.org/r/20221025191339.667614-3-helgaas@kernel.org
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>