phy: qcom-qmp-pcie: split pcs_misc init cfg for ipq8074 pcs table
Commit af6643242d3a reworked the pcs regs values and removed the
0x400 offset for each regs.

This change caused the malfunction of ipq8074 since it still has the
legacy pcs table where pcs_misc are not placed on a different table and
instead put together assuming the offset of 0x400 for the related
pcs_misc regs.

Split pcs_misc init cfg from the ipq8074 pcs init table to be handled
correctly as they are at an offset of 0x400 and add ipq8074 to the
workaround list used for similar device that suffer from the same
implementation like ipq6018.

The workaround is needed to save on compatibility with old DT where
pcs_misc is not declared and instead is part of the pcs.

Fixes: af6643242d3a ("phy: qcom-qmp-pcie: split pcs_misc region for ipq6018 pcie gen3")
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>