gfs2: Cease delete work during unmount
Add a check to delete_work_func() so that it quits when it finds that
the filesystem is no longer writeable.  This speeds up the delete
workqueue draining in gfs2_kill_sb().

In addition, check for the SDF_READONLY flag before queuing new delete
work in iopen_go_aux_work(); otherwise, new work could get queued after
we have flushed the delete workqueue.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>