gfs2: Make glock lru list scanning safer
In __gfs2_glock_put(), remove the glock from the lru list *after*
marking it dead and dropping the glock lock.  This prevents lock
inversion against gfs2_scan_glock_lru().

In gfs2_scan_glock_lru(), make sure that the glock's reference count is
zero when moving the glock to the dispose list.  This skips glocks that
are marked dead as well as glocks that are still in use.  Additionally,
switch to spin_trylock() as we already do in gfs2_dispose_glock_lru();
this alone would also be enough to prevent lock inversion against
__gfs2_glock_put().

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>