ath11k: support setting FW memory mode via DT
ath11k is really memory intensive for devices with less that 1GB of RAM,
so lets allow saving a significant amount of memory by setting the FW to
Mode-1 via DTS for devices that need it.

However the drawback is reduced number of VDEV-s and peers which is a
reasonable tradeoff.

Mode-2 allows for further reduction, but it has further restrictions.

While we are here, lets add a print to be able to easily determine what
FW memory mode is being used.

Signed-off-by: Robert Marko <robimarko@gmail.com>