dt-bindings: cpufreq: qcom-cpufreq-nvmem: make cpr bindings optional
The qcom-cpufreq-nvmem driver supports 2 kind of devices:
- pre-cpr that doesn't have power-domains and base everything on nvmem
  cells and multiple named microvolt bindings
- cpr-based that require power-domain in the cpu nodes and use various
  source to decide the correct voltage and freq

When the schema was introduced, it was wrongly set to always require these
binding but this is not the case for pre-cpr devices.

Make the power-domain optional and set them required only for qcs404
based devices.

While at it also make more clear what the opp-table supports by adding
ref to the opp-v2-kryo-cpu and opp-v2-qcom-level schema.

Fixes: ec24d1d55469 ("dt-bindings: opp: Convert qcom-nvmem-cpufreq to DT schema")
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>