arm64: dts: qcom: ipq6018: add thermal nodes
IPQ6018 has a tsens v2.3.1 peripheral which monitors temperatures around
the various subsystems on the die.

We can use those temperature sensors in order to protect the SoC from
overheating, so lets setup passive cooling trips to cool the CPU via
CPUFreq downlocking it, and if that doesnt work trigger the criticla trip.

Signed-off-by: Robert Marko <robimarko@gmail.com>