net: phy: aquantia: enable USXGMII autoneg
In case USXGMII is being used as the PHY interface mode then USXGMII
autoneg must be enabled as well.

HW defaults to USXGMII autoneg being disabled which then results in
autoneg timeout, so enable it in case USXGMII is used.

Since AQR-s store the actual PHY mode being used in an register, make
sure that is being parsed after FW comes out of reset first so the
interface mode in phy_device gets updated as well.

Signed-off-by: Robert Marko <robimarko@gmail.com>