net: phy: aquantia: drop wrong endianness conversion for addr and CRC
On further testing on BE target with kernel test robot, it was notice
that the endianness conversion for addr and CRC in fw_load_memory was
wrong actually not needed. Values in define doesn't get converted and
are passed as is and hardcoded values are already in what the PHY
require that is LE.

Also drop the cpu_to_be32 for CRC calculation as it's wrong and use
_swab32 instead, the word is taked from firmware and is always LE, the
mailbox will emi a BE crc hence the word needs to be always swapped and
the endianness of the host needs to be ignored.

Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202311210414.sEJZjlcD-lkp@intel.com/
Fixes: e93984ebc1c8 ("net: phy: aquantia: add firmware load support")
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>