remoteproc: qcom_q6v5_mpd: add support for passing v1 bootargs
On multi-PD platforms such as IPQ5018, boot args are passed to the root PD
run on the Q6 processor which in turn boots the user PDs for internal
(IPQ5018) and external wifi radios (such as QCN6122). These boot args
let the user PD process know details like what PCIE index, user PD ID, and
reset GPIO is used. These are otherwise hardcoded in the firmware.

Below is the structure expected of the version 1 boot args including the
default values hardcoded in the firmware for IPQ5018:

+------------+------+--------------+--------------+
| Argument   | type | def val UPD2 | def val UPD3 |
+------------+------+--------------+--------------+
| PCIE Index | u32  | 0x02 (PCIE1) | 0x01 (PCIE0) |
| Length     | u32  | 0x04         | 0x04         |
| User PD ID | u32  | 0x02         | 0x03         |
| Reset GPIO | u32  | 0x12         | 0x0f         |
| Reserved 1 | u32  | 0x00         | 0x00         |
| Reserved 2 | u32  | 0x00         | 0x00         |
+------------+------+--------------+--------------+

On IPQ5018/QCN6122 boards, the default mapping is as follows:

          +-> UPD1 ----> IPQ5018 Internal 2.4G Radio
         /
        /
Root PD +---> UPD2 ----> QCN6122 6G Radio on PCIE1 (if available)
        \
         \
          +-> UPD3 ----> QCN6102 5G Radio on PCIE0

To support (future) boards with other mappings or control what UPD ID is
used, let's add support for passing boot args for more flexibility.

Signed-off-by: George Moussalem <george.moussalem@outlook.com>