net: ethernet: qualcomm: Initialize PPE L2 bridge settings
Configure the default L2 bridge settings for the PPE ports to
enable L2 frame forwarding between CPU port and PPE Ethernet
ports.

The per-port L2 bridge settings are initialized as follows:
For PPE CPU port, the PPE bridge TX is enabled and FDB learn is
disabled. For PPE physical port, the PPE bridge TX is disabled
and FDB learn is enabled by default and the L2 forward action
is initialized as forward to CPU port.

Signed-off-by: Lei Wei <quic_leiwei@quicinc.com>
Signed-off-by: Luo Jie <quic_luoj@quicinc.com>