spi: spi-qpic: fix broken driver with SPINAND_SET_FEATURE command
The driver always return probe error with SPINAND_SET_FEATURE command:

spi-nand: probe of spi0.0 failed with error -1207959538

The error doesn't match any expected negative error but instead seems to
be an u32 converted to an int. Investigating the entire codeflow I
reached the culprit: qcom_spi_cmd_mapping.

Such function can return -EOPNOTSUPP or the cmd to run. Problem is that
in the specific context of SPINAND_SET_FEATURE, BIT(31) is set that in
the context of an integer, it gets treated as a negative value.

To correctly handle this, rework the function to return 0 or a "correct"
negative error and pass a pointer to  store the cmd.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>