mips: replace -mlong-calls with -mno-long-calls if possible
This is a really old patch ported from MikroTik. It needs a an
additional patch to actually be implemented and both this and the old
one are considered HACK as they bypass normal kernel linux to make it
work.

The original message quote:

replace -mlong-calls with -mno-long-calls to make function
calls faster in kernel modules to achieve this, try to load
kernel modules to KSEG0 and if that doesn't work, use vmalloc
and fix up relocations with a jump table based on code from a
kernel patch by MikroTik.

SVN-Revision: 16772

lede-commit: 3b3d64743ba2a874df9d70cd19e242205b0a788c
Signed-off-by: Felix Fietkau <nbd@nbd.name>
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>