rtc: rs5c372: let the alarm to be used as wakeup source
Currently there is no use for the interrupts on the rs5c372 RTC and the
wakealarm isn't enabled. There are some devices like NASes which use this
RTC to wake up from the power off state when the INTR pin is activated by
the alarm clock.

Enable the alarm and let to be used as a wakeup source.

Tested on a Buffalo LS421DE NAS.

Signed-off-by: Daniel González Cabanelas <dgcbueu@gmail.com>